Implement idle Unit reaping
Summary:
Right now when a Unit is created, it is only ever deleted if the path
it is associated with gets a new Unit. If changes are done via symlink
changes (thus creating new paths), or if a Unit is simply not accessed
frequently, it will stay around forever.

Implement a facility to "reap" idle Units. If Eval.IdleUnitTimeoutSecs
is greater than zero, any Unit which has not been used by a request in
more than the configured seconds will be
deleted. Eval.IdleUnitMinThreshold is an optional configuration which
will prevent the Unit reaper from dropping the total number of Units
below the configured amount. This is useful to prevent all of the
Units in an idle server from being freed (it can maintain a core
"working set").

Add two new pieces of data to Unit: The first is the newest request
which has used that Unit. This is updated whenever a Unit is loaded by
a request. The second is a timestamp for the latest time the Unit was
used by a request. This is updated by the request when the request is
about to end. To reason for deferring the timestamp update until the
end (rather than when it is loaded) is to avoid biasing against long
running requests. In the extreme, a long running request which touches
all of its Units early could have then reaped immediately when it
ends. So, instead the expiration is measured from the end of the
latest request which used it.

The Unit reaper itself is just a thread which periodically wakes up,
scans all of the Unit caches (both path and eval caches) and looks for
any Units which are expired. A Unit is expired if its request stamp is
for a non-running request (this means no request is currently using
it), and if its timestamp plus expiration time has already
past. Expired Units are removed from their respective caches and then
put on the treadmill to be eventually deleted.

Reviewed By: paulbiss

Differential Revision: D24385329

fbshipit-source-id: f7ddd827170fed250a8c6eacff51fe48d1733dc9