Make profiling translations only guard to DataTypeSpecific for arrays
Summary:
This diff has been reworked from the original version. We take the following approach:

Profiling tracelets guard only to DataTypeSpecific for arrays. When we perform a layout-sensitive operation, we first record that an array reached the layout-sensitive bytecode in that tracelet. We then guard to vanilla. If the guard fails, we guard to logging. If the guard fails, we side-exit. Vanilla and logging implementations are left unchanged.

When we emit live tracelets, we preserve the bespoke layout of the array in our live types and constrain to DataTypeSpecialized when we reach the usage site.

When we emit optimized tracelets, we will use coloring information to determine which layout to guard to before the layout-sensitive operation.

Differential Revision: D24307233

fbshipit-source-id: 376c787020af34fe80561d0b9a09b2e1c12856ad