Deflake unit-cache-hash Unit test
Summary:
I've noticed this test (very rarely) failing spuriously. One
of the things it tries to do is create a Unit with a different SHA1,
but identical HHAS (to exercise the reuse logic in the unit-cache). It
seems that there's some non-determinism in HackC which means that
occasionally the HHAS will be different, causing the test to fail (the
Unit is not reused). Remove this portion of the test

Reviewed By: paulbiss

Differential Revision: D24413759

fbshipit-source-id: 1994bfdf73493e18470c73cef30505722630568c