C++ize bcmath extension
Summary:
I want to fix a number of issues in the bcmath extension, but that
will require dealing with exceptions and other C++ things. Since
bcmath is currently C code, compile it as C++. Rename the file
extension and remove the ancient C function declaration syntax that
isn't supported in C++. I'm splitting these changes (which are
strictly mechanical) out to make the actual improvements easier to
review.

Reviewed By: oulgen

Differential Revision: D24402101

fbshipit-source-id: 69bd521cb3291d2b225b71395bd3d7b43a7302c2