Fix issues in bcmath
Summary:
bcmatch has the "interesting" behavior that when it thinks it runs out
of memory, it just calls exit(). This is obviously bad for a
server. What's even worse is that this behavior can be easily
triggered by inputs to the various bcmath functions (just pass in a
huge scale). So a Hack program can easily just cause the server to
exit. We also don't have any memory controls on the extension, so its
easy to make it allocate huge amounts of memory (far greater than the
memory limits).

Fix these issues. Replace raw calls to malloc() to use bc_malloc(),
which is a wrapper which calls the request malloc functions. So all of
bcmath's allocations will come out of the request heap and be
accounted for. We do OOM checks as well. Replace the "malloc_size"
check in init.cpp to do the check from within bc_malloc() (it's not
clear if this check is actually needed. The diff which introduced it
didn't sound confident). Since bc_malloc() can now throw, use
SCOPE_EXIT and SCOPE_FAIL to provide some moderate degree of exception
safety in the C code (I do not pretend that this is 100% complete).

Delete "outofmem.cpp" since we don't want to exit when we run out of
memory. Since bcmatch now expects all numbers to be in the request
heap, rework the RDS local logic that initializes a few constants. As
an added benefit, we don't pay any cost if we don't use the extension,
and the "precision" setting doesn't carry over between requests.

The extension would also sometimes print to stderr. Redirect it to
emit warnings or recoverable errors instead.

Add unit tests which trigger some previous issues.

Reviewed By: oulgen

Differential Revision: D24406551

fbshipit-source-id: 6f9fda476d57fb899d34c5a8ac199f957878cfc1