Allow constraint solver to be position blind
Summary:
The constraint deduplicator in the simplifier was not doing its job very well because during the solving we also carry position sets around. So we did not deduplicate unless two flows had exactly the same position sets. This is not necessary for correctness, so this diff makes the deduplicator position set insensitive. As such there are fewer constraints to eliminate.

In particular, this prevents exponential behaviour in the following code:
```
$d[4][3][2] = 0;
$d[4][3][2] = 0;
$d[4][3][2] = 0;
$d[4][3][2] = 0;
$d[4][3][2] = 0;
$d[4][3][2] = 0;
```
Each assignment can throw, so they add to the pc set and we get an extra but identical (except for the position) flow despite analysing identical code. During existential quantification elimination, if we have `N` replica of this constraint and `M` flows that can combine with it, we get `N * M` constraints rather than just `M` constraints.

The implication of this to the error reporting is that if there are multiple illegal flows from `A` to `B` in the same function we only report one because we do not distinguish them.

Reviewed By: mpu

Differential Revision: D24308275

fbshipit-source-id: 52871bef78dcf979721b118cf14c15daa7e49ab3