Add support for detecting cooperlake CPUs
Summary:
We detect cpu models and conditionally enable optimizations based on this.
Cooperlake appeared to be improperly detected.  I also make note that the
current logic that assumes that monotonically increasing model numbers is
wrong.  Rather than fix it immediately, I left it so we don't change behavior
on exisiting systems for now.

Reviewed By: ricklavoie, meteorfox

Differential Revision: D24450678

fbshipit-source-id: d705fed05a554c5aa164dfca20b0d0edd1c55f81