Add param provenance for variadics
Summary:
We add a new kind of array provenance, which points to a callee and its argument, to be used for variadics and intercepted functions - basically, cases where the runtime constructs an array as part of some edge case in the calling convention.

We use expensive string formatting to stuff a bunch of new information into this tag. Since it's only constructed in the interpreter and the JIT, and it's burned into JIT-ed code, there's no performance issue.

Reviewed By: dneiter

Differential Revision: D24436473

fbshipit-source-id: d3803d29e20ec5ac5fabd2edceffd0b4b8442db7