Perform lval updates inside Elem operations
Summary:
Elem{Int,Str} functions now take a tv_lval as a parameter instead of an ArrayData. They are responsible for making any updates to the lval should escalation or copying be performed. Further, a parameter indicating whether a missing element should throw or not is provided. Usage sites in member-operations and implementations are updated.

This will aid in JITing Elem operations for bespokes.

Differential Revision: D24428587

fbshipit-source-id: b7ea1e616fab0d05d492716b905d3c4846b3ceed