Stop using FuncId as an index in profiling code
Summary:
In LowPtr builds, I'm planning to convert FuncId into Func* pointer itself. This allow us to kill s_funcVec in LowPtr builds.

In order to do this, I need to remove all use cases of FuncId as indices into a vector.

Note that this diff does not increase the space consumption of HPHP::Func since the storage space was currently wasted due to alignment.

Reviewed By: jano

Differential Revision: D24371160

fbshipit-source-id: ecaa8190d641401627af4f81f174a5c6eef80f65