Fix diamond behavior for vanilla builtins
Summary: Some builtins (and the IR ops that invoke them) are known to return vanilla arrays. Do not attempt to emit a diamond if we have a known vanilla array, or we'll wind up with a bottom type on the bespoke side of the diamond. Introduce a new test to check this case.

Reviewed By: mofarrell

Differential Revision: D24476723

fbshipit-source-id: 85198d2bff67551d6ee9805b176b6860f409da46