SrcKey: clean up and generalize prologue bit logic
Summary:
Prepare the field that stores the resume mode and the prologue tag for addition
of the func entry tag. Change its name, define constants specifying the number
of used bits for this field and the offset and update assertions (the were
using out of date numbers) to use these constants.

Reviewed By: mofarrell

Differential Revision: D24440366

fbshipit-source-id: cb82886d6784d777200c08594bec10b57a0848b7