SrcKey: implement lineNumber()
Summary:
There are many places in the codebase that try to obtain a line number for
a SrcKey.

Move this functionality to SrcKey::lineNumber() and unify it so that prologues
point to the first line of the function declaration.

While there, fix two bugs:
- vtune-jit.cpp: resolving line number from a line number (my recent bug)
- print.cpp: iroff() is a sequence number of IR instruction within a bytecode, adding it to bcOff() is meaningless and computing a line number from that even more so; the rest will be fixed in the next diff

Eliminates some usage of offset()s in prologues and improves debugging and
tracing output.

Reviewed By: oulgen

Differential Revision: D24441375

fbshipit-source-id: 4a7675b7b6aec94512456bbc59b0e3a070f2f75a