Codemod IFC to use Naming_special_names instead of artisinal attributes
Summary:
Since we've now defined all the special attributes in Naming_special_names, we can remove them from the custom decl phase, as well as get rid of the custom magic attribute classes they needed. This diff does not *consume* the declarations implemented in the diffs below it, but moves the references to names to a central place.

It also renames Governed to Policied, to be consistent everywhere.
This was done via a fastmod script:

```
fastmod -F "<<Policied" "<<__Policied" -g "hphp/hack/test/ifc/**/*.php"
```
and similar ones for InferFlows and Governed.

Reviewed By: zilberstein, mpu

Differential Revision: D24318131

fbshipit-source-id: 8ec927a73feb1d3ac3250258bf5f660c0257a797