inline Decl_provider.get_class_constructor
Summary:
This diff gets rid of a few Decl_provider APIs:
* get_class_constructor - short hand for getting a class, then getting its constructor, then converting it into a fun_decl
* get_class_method - as above but for methods
* get_static_method - as above but for static methods

The three functions were introduced in D16866684 (https://github.com/facebook/hhvm/commit/4da2248506c69b3cb55053af7af661902726c548), and are used only once in the codebase, in D16867830 (https://github.com/facebook/hhvm/commit/1079b0142580354d88856e511f1c5371bf108178) for signature help.

I think the're not good APIs: (1) we already have a good API in our decl surface area, in which first you get a class then you get a method on it, and we shouldn't be in the business of providing compound accessors; (2) the idea of converting a class_elt to a fun_decl doesn't make sense, and is being done by the consuming code here solely as a shortcut so that common elements of both can be displayed.

Therefore: I'm removing the functions, and inlining them in their callsite.

The result of this diff is that Decl_provider has a very clean, streamlined API. It solely has public methods for the top level entities (classes, functions, typedefs, gconsts, record-defs), and it implements each of them in a nice uniform way.

Differential Revision: D24435235

fbshipit-source-id: f81672861a198f6f0415b40ea4004678e5f371f7