thread count_decl context through to count_subdecl
Summary:
My overall goal is to understand which decl-accessors are invoked, and how frequently.

What this diff adds is the ability, when recording telemetry about accessors like "does this class implement IDisposable", to also correlate that back to the telemetry of the earlier "get class".

We'll see it used in practice further up the stack.

For the current diff, my concern is solely to (1) add this possibility, (2) have it be as performant as possible: it should add no overhead to the production case where we're not gathering telemetry about decl-accessors.

Towards the performance end, I've made a slight telemetry change. It used to be that, in prod, we were counting every single decl accessor (e.g. get_method). I don't think that was worth doing in prod. Sure it's nice for us to do analysis and get insight. But for prod, I think just the top level data of how many funs+classes+records did we fetch is enough.

Reviewed By: CatherineGasnier

Differential Revision: D24385207

fbshipit-source-id: 313fb378aabf03b6208482bd3550bb0d7cefad4f