inline Typing_lazy_heap into Decl_provider
Summary:
I'm a firm believer in "equal yokefellows" http://www.171english.cn/html/kingenglish/503.html

In this case, Decl_provider has
1. a "Sharedmem_provider" which calls into Typing_lazy_heap to do caching around a call to the underlying Decl.declare_fun_in_file
2. a "Local_provider" which, inline, itself does caching arounda  call to the underlying Decl.declare_fun_in_file.

Both providers are doing equivalent things. Therefore they should be doing it in the same way, either both inline or both in a separate function.

I've chosen to do it inline because the functions still remain adequately readable, and it's better to see similarities+differences between the two calls. For instance, bringing it inline has made manifest that we're using the `Counters` module wrong. I'll address that in the next diff.

Separately: I always disliked the name "Typing_lazy_heap". I don't think that was a good description of what it does, so I'm happy to get rid of it. I think that "Decl_provider" is a much better name for this functionality.

Differential Revision: D24434817

fbshipit-source-id: c1c4c91cccad33e88a72e87a49168d5215c3326a