avoid re-entrant calls to Counters.ml
Summary:
We're not counting decls right in all cases. Throughout this diff stack I'm trying to tighten up on it.

My goal is to know precisely which decl-accessor gets invoked by external code, and how many times. In other words I want a strong layer of abstraction around decls, a kind of force-shield. And I want to know precisely how many times someone crosses that force-shield.

(Why? Because my end goal is to figure out perf of the shared decl service, and to figure out which decl APIs need be shared inside it to improve perf. The force shield will ultimately correspond to the API of the decl service).

Here's a concrete example. There are several entry points that are called by consumer code:
* Classes.get
* Classes.mem
* Classes.find_unsafe

Prior to this diff, if someone called Classes.mem, then it actually got logged TWICE under the decl-accessor counter, once for Classes.mem and once for the inner call it makes to Classes.get.

With this diff I'm removed those double-counting cases. Now, we only count at times when an external user invokes the public API.

Reviewed By: CatherineGasnier

Differential Revision: D24384723

fbshipit-source-id: 148d010bee9c6f75e14c9ea352e4298994c416a5