Simplify per-ArrayKind sort helpers
Summary:
A lot of the logic in EscalateForSort for different array kinds is no longer needed.

At the same time, I want to avoid calling EscalateForSort in cases where a) the array is empty and b) no vec->dict escalation is needed. I modified array_sort helpers to return early in these cases. All this is in preparation for a minimal-power BespokeArray::EscalateForSort. As an invariant: I never want to call EscalateForSort on an empty array except for vec->dict escalation.

Differential Revision: D24471056

fbshipit-source-id: 013de19fb972d7ea8b9b9aceaee78c059c719921