Fix union and intersection type formatting
Summary:
hackfmt's treatment of union and intersection types has always been wrong. We used the `transform_argish` helper here, but it's not the right choice for this kind of separated list--it's for comma-separated lists, and adds trailing commas by default.

This diff avoids using the `transform_argish` helper, producing the correct formatting.

Reviewed By: mjhostet

Differential Revision: D24461232

fbshipit-source-id: 234611cdce229aff793833af2dd67157e876fb77