SrcKey: split offset() from entryOffset()
Summary:
SrcKey::offset() is assumed to return a bytecode offset corresponding to the
SrcKey location. This is obviously not true for prologue locations, where the
offset() happens to return the offset of the function entry. It is easy to
misuse offset() and try to interpret it as a location of the current bytecode.

Split this function into offset() and entryOffset() and add the appropriate
assertions to ban incorrect usage.

Audit places where these APIs are used and make sure they do the right thing.
Unsurprisingly, this was mostly limited to debugging and tracing output.

Reviewed By: paulbiss

Differential Revision: D24440600

fbshipit-source-id: ae461a59e90e3a9a63cbf955836938d6c446b66f