Fix updateMarker() in prologues by using SrcKey instead of Offset
Summary:
The prologue tag is dropped from the current marker's SrcKey when updateMarker()
is used. This is because makeMarker() accepts an Offset and uses it to build
the new SrcKey.

Fix this by accepting a SrcKey instead of an Offset. Update some APIs that
create a marker (such as makeExit()) to do the same. These APIs were usually
passed bcOff(env) or nextBcOff(env), which were simply replaced by
curSrcKey(env) or nextSrcKey(env), preventing a SrcKey -> Offset -> SrcKey
roundtrip.

Reviewed By: paulbiss

Differential Revision: D24470012

fbshipit-source-id: 1720ecd58fda64b5ef92d89906494702ddd11e2d