Add position metadata to desugaring
Summary:
`ExprTreeVisitor`s should be able to access the positional information about the constructs that they are visiting. This is useful for reporting errors and may be useful for caching purposes.

When desugaring, emit a `ExprPos` object containing the starting and ending line and column number.

In the case of a `Pos.none`, we construct a `ExprPos` with 0s.

Reviewed By: Wilfred

Differential Revision: D24481995

fbshipit-source-id: b7af45ec8c1392c23e04d2fed6bfa038ec16b36e