Move test definitions to bottom of file
Summary:
Everytime we change the definitions of one of the APIs, it changes all of the positions of the errors that are reported.

I want to move all of the API definitions below the code actually being tested to avoid needed to reupdate all of the error positions.

I also added the following lines
```
//// BEGIN DEFS
```
and
```
//// END DEFS
```
To act as little guides so that the script in D24466756 can do it's work. The script allows you to update one file's defs and change all of the other files at the same time. Since they are comments, I figure it's mostly harmless.

I'd like to commit this even if we don't want to commit the script that does the update.

Reviewed By: Wilfred

Differential Revision: D24466762

fbshipit-source-id: 67e08f6ab0dd388e28c8db3190d1fbce537d1540