Kill a bunch of unnecessary makeCatchSet blocks
Summary: We only need to use a `makeCatchSet` block if the helper we're calling can throw InvalidSetMException. This only occurs when the base of the SetM op is a scalar, a string, or null. Therefore, if we have enough knowledge of the type to conclude it is impossible to throw InvalidSetMException, don't create a custom catch block.

Differential Revision: D24504732

fbshipit-source-id: b868f6ba3cb23524b2a4dcccf889aa37d8b5af60