Make jit/interp null/fcfp comparison consistent
Summary:
There is a difference in the behavior of comparison (`<`, `>`, etc.) for fcfps and other data types between jit and interp mode. For a couple of cases, the error message differs.

But for null, resource, and some other types, the difference is between throwing an error and returning an actual value. This diff fixes the latter. Now comparison with null, resource, obj will always throw errors in jit, matching up with interp behavior.

Reviewed By: paulbiss

Differential Revision: D24234317

fbshipit-source-id: 614f2889a95a3edd324212463e8b682b8a377329