Tweak irgen for bespoke layout checks
Summary:
This change is a rework of how we guard in irgen-bespoke. The goal is to make the nontrivial cases more clear - to do that, we fold the trivial cases (which are all very similar) into a getLocationToGuard helper and reduce the amount of work done by things like emitLoggingDiamond.

The other change here is that now all layout checks are done with a base type of TArrLike, so that we don't care whether we previously guarded the given location to a specific array-like type or to TArrLike alone. This change is a precursor to updating iterator ops to have looser guards. To make this change performant, we update cgCheckType to skip the DataType checks for CheckType<ArrLike=Vanilla> on a TArrLike input.

When we update iterator bytecodes, we now only need to change two places - getLocationToGuard will use a looser GuardCategory for iterator bytecodes, and we'll move iterators to the first dispatch case in handleBespokeInputs. The reason I pulled emitLogArrayReach out of handleBespokeInputs is because it applies to both the iterator and the diamond cases.

We handle the bugs fixed by D24476723 (https://github.com/facebook/hhvm/commit/cb3376dca178945c1617317479365f4ed00ea76b) and D24479155 in this diff - for the first one, that's our first dispatch case, and for the second, we do an early bail-out.

Differential Revision: D24479194

fbshipit-source-id: 8b6d24ea254dd6edcf23013dd3ec1a692f80a2f4