Merge array-like iterator tracelets
Summary:
If an iterator is used for multiple array-like bases, it's better for us to just guard to TArrLike and emit a native call. Currently, we guard to each distinct array-like type and emit the same native call for each one.

We do so by introducing a new DataTypeIterBase category, and by ensuring that iterators are in tracelets of their own (so that we don't strengthen this guard for any other op, like the UnsetL that commonly follow IterNext bytecodes). One small wrinkle is that we have to avoid encoding the DataTypeSpecific values in the type post-conditions of the tracelet. That's because if we end up doing those, then we won't relax the LIterInit and LIterNext bytecodes for different array-like types to be a single optimized tracelet - they'll chain into the next tracelet, and we can only merge tracelets with identical successor sets.

An alternative design here is to just use DataTypeGeneric plus manual checks here. The reason I prefer this approach is because iteration is genuinely different for array-likes and objects. We could maintain this difference by emitting profiling code and then optimizing to one or the other eventually, but that's just rebuilding the guards system by hand.

See here for more info: https://fburl.com/pxxn9xvs

Reviewed By: ottoni

Differential Revision: D23725861

fbshipit-source-id: 29c2612e3196b9bd6b7950ba91a282848b7a1a7b