Group logs by (Func, Offset) alone
Summary: For the purposes of bespoke logging, we don't need to distinguish between resume modes or by the other new fields we're adding to SrcKey for the calling convention work. We do so by "canonicalizing" our SrcKey to be a (Func, Offset) alone. That happens in logging-profile so that we can change granularity if we want.

Reviewed By: ricklavoie

Differential Revision: D24528413

fbshipit-source-id: edf57561c1a6b20a1bae45f169b7447f94d2935d