Instrument casts from the interpreter
Summary:
I missed this case. We want to instrument all static array ops, all new array allocation ops, and all casts.

We've already instrumented {static array, array allocation, cast} from the JIT and {static array, array allocation} from the interpreter, which are way hotter than casts from the interpreter. There should be no perf effect.

One tricky thing here was that we were previously trying to avoid treating a vec->vec or a varray->vec cast as a new array constructor. We did so by checking whether the output of the cast is known to be vanilla in the JIT, which is true for layout-changing casts but not for these simple cases. We're going to replace this approach with a "sampled" bit in the next diff, so we can drop those hacks.

Reviewed By: ricklavoie

Differential Revision: D24528987

fbshipit-source-id: 112cdd7504ab8894c2c0b75c9e0cd0968020fbd7