Add sampled bit to arrays
Summary:
We have full LoggingArray coverage for "Hack array creation sites" (which comprise static array constructors, like Arr/Vec/Dict/Keyset, new array allocations, like NewVec/NewStructDict, and casts).

However, we have not instrumented array constructors in builtins, of which the most important are deserialization functions (e.g. json_decode and Thrift). If we use LoggingArray data to determine that a sink only sees monotype vecs and specialize it for that case, but it also gets arrays from builtins, we could significantly hurt performance. mcolavita and I call this problem a "probability leak".

We solve this problem by adding a new isSampledArray() bit to vanilla array-likes. When we decide not to create a LoggingArray for a given vanilla array in maybeMakeLoggingArray, we'll set this bit instead. If we see any unsampled vanilla arrays at a given sink, we must have a probability leak.

Reviewed By: ricklavoie

Differential Revision: D24529082

fbshipit-source-id: 6c84f6ab4c0a94d14b17996ae56039910e7e39e9