Pass LoggingProfile to the NewLoggingArray call
Summary:
In the JIT, we can burn this profile in as an immediate rather than doing a concurrent hash map lookup every time. In addition to saving on lock contention that way, it's also good for code quality - we now centralize the logic for "should I profile this bytecode?" in one place, in getLoggingProfile.

This kind of change will matter more for SinkProfile, which I'm adding next, because we average 10-100 sink operations for each sampled array.

Reviewed By: ricklavoie

Differential Revision: D24531812

fbshipit-source-id: 4a925a769f4b2885b4f89e28d2f5b5946904e1b7