Type constant access through intersections and unions
Summary:
This is an improved treatment of type constant access in the presence of unions and intersections, necessary for soundness in complex cases.

Typically we access type constants through a bounded generic e.g. `T as B` and then `T::TP` in a `where` constraint. What if `B` is an intersection or union? This isn't denotable except under a flag, but can come about through flow (unions) and multiple bounds or refinement (intersections). A union is only safe if the type constant is present in *every* disjunct in the union; dually, an intersection is safe if the type constant is present in *any* conjunct. Bounds on type parameters are combined pointwise.

So, for example, if we have
```
interface I1 {
  abstract const type TP as J;
}
interface I2 {
  abstract const type TP as K;
  public function get():this::TP;
}
interface I3 {
  public function another_method():void;
}
```
then it's legal to allow access to `this::TP` through `T as I1 as I2 & I3`, with resultant bound `J & K`. Or  suppose we add
```
interface I4 {
  abstract const type TP as L;
}
```
and `T as I2 & (I1 | I4)`. Then it's legal to access `this::TP` but its bound should be treated as `K & (J | L)`.

The code is hopefully cleaner, without any use of exceptions or backtracking "run_on_intersection".

Note that the use of `global_tpenv` is still broken (and explains the conservative treatment of abstract type constants in `union_results` and `intersect_results`). Hopefully this will be superseded in a future diff by a new treatment of type constants.

Reviewed By: dlreeves

Differential Revision: D24358865

fbshipit-source-id: e0abcc790225f8c2d889be286065d8e4c81ed6bb