Kill TransFlags
Summary:
This was used a while ago as a mechanism to side-exit when encountering
unlikely calling convention situations that would otherwise require a lot
of code to be emitted to be handled properly.

The code bloat issues were resolved by improving the language semantics and
the calling convention implementation. This mechanism is not used for over
a year, so let's kill it.

Allows further improvements in the service request area.

Reviewed By: ricklavoie

Differential Revision: D24528554

fbshipit-source-id: 39564188a48a66c95fd093ee086b8f5d1332fb30