Fix another APCArray arrprov bug
Summary:
APCArray should get a provenance tag iff the array we're trying to store gets one. Breaking this invariant causes us to crash, because we store the tag inline with the APCArrray allocation - if we get it wrong, we'll return misaligned allocations to jemalloc. In particular, we'll trigger this bug if we apc_store and then apc_delete a marked dvarray.

(Note that the if the array transitively contains pure value types - i.e. only scalars or other arrays - we'll turn the array into a static or uncounted array and avoid creating an APCArray at all. That's why we don't crash a million times over in prod.)

Reviewed By: dneiter

Differential Revision: D24552992

fbshipit-source-id: 8e4e1332771d7328a6d238f456314b200fcbac29