One more varargs runtime location
Summary:
Like all other runtime locations that construct a varargs varray as part of our calling convention, this location should tag the new array as being a function's variadic param.

dneiter: note that if the caller is doing any kind of unpacking call, provenance will be attributed to the caller, not the callee. That could pose a problem for cases where the number of unpacked arguments doesn't match the final number of varargs. Let me know if you want it changed - it's won't be that hard to do so.

Reviewed By: dneiter

Differential Revision: D24555348

fbshipit-source-id: 81d324deb23aecaae891f552b198fcc60d09d1b2