Fix type updating array assignment bug
Summary:
When the assignment to a Hack array causes the type to change, e.g.,
```
function f(vec<int> $v, string $s) {
  $v[] = s;
}
```
we were assuming the type must be what it was before leading to erroneous subtype calls that fails.

This diff fixes this by making the old type of `$v` a subtype of the type _after_ the assignment which is conveniently stored in the LHS of the assignment.

Reviewed By: mpu

Differential Revision: D24505198

fbshipit-source-id: 33cf19115cd0c7f05704f2bd71fdb749e1f1cb95