Raise error when a class inherits multiple methods differing only by case
Summary:
This diff implements a tast check which, for every class, checks that its members do not conflict only on case, e.g. foo vs. FOO.

HHVM is case insensitive when it comes to method calls (that is, $x->foo is always the same as $x->FOO), so these are technically safe today, but HHVM is working to make method calls case sensitive. Thus, this behavior has to be banned ahead of time in order not to create large unintended errors in the codebase, where it is guaranteed that foo vs. FOO mean the same thing.

Reviewed By: Wilfred

Differential Revision: D24149745

fbshipit-source-id: 057e591d4d509c439acfbb0fdbe1440349a81587