Add Shallow_decl_defs.decl type
Summary: I think it would be easier to understand the output of the direct decl parser tests if decls were presented in the same order they appear in in the file. This is difficult to do with the current structure of the `decls` type, which partitions decls into separate maps for functions, classes, typedefs, records, and constants. Instead, I'm suggesting that we use a new type `Shallow_decl_defs.decl`, which is a variant type containing all five kinds of decl, and represent the collection of decls parsed from a file with a `(string * decl) list`.

Reviewed By: shiqicao

Differential Revision: D24549317

fbshipit-source-id: bb9aa0ae8c45d1a861fe0bdd1f1141e3b1abfde5