Simplify hh_single_decl
Summary:
This test runner for the direct decl parser is a little more complex than it needs to be. It discovers decl names in the file under test using the facts parser, then uses the legacy decl pipeline to produce a decl for each name, and the direct decl parser to produce decls for the entire file. For each decl, it compares the legacy value and the direct decl value using derived `eq` functions. If they differ, it prints a string diff between the `show` representations of the two decls.

This leads to strange situations when `eq`'s notion of equality differs from `show`-output equality (for instance, `eq` ignores typing reasons).

I think it's simpler if we just use `show`-equality here. Then we can simply parse and declare the file using the legacy pipeline and compare the resulting decl list to the output of the direct decl parser (instead of fussing with the facts parser or abstracting over five different kinds of decl in order to perform per-decl comparisons). This change exposes that 11 of our test cases were passing incorrectly (and now, no longer print "They matched").

Reviewed By: shiqicao

Differential Revision: D24549371

fbshipit-source-id: fc4d6b008ecca3ff1ccf70e22f76f1408da17233