Make Taccess both a locl_ty and a decl_ty
Summary:
`Taccess` is used to represent decl types such as `this::T1::T2` and `C::T` and `T1::T2` (in where clauses). It's localized by expansion to a concrete definition, or through generation of mangled fake generic parameters with bounds in `global_tpenv`.

In future we will not always elaborate type constant access during localization, instead treating `::T` as a type operator in its own right, with subtype rules, etc.

This diff is a first step: allow `Taccess` to be a locl type.

Reviewed By: vsiles

Differential Revision: D24576802

fbshipit-source-id: 7e57a41aef15f41e28e1f9be82871593ee582962