Track the position of fixed-width tokens
Summary:
For performance and memory usage reasons, we'd like to avoid performing allocations in the direct decl parser as often as possible. As a result, we've distinguished tokens whose positions we need to track from tokens whose positions we don't need to track--we represent the latter with Node::Token(TokenKind), and we represent the former with a variety of Node variants (Node::Backslash, Node::Operator, Node::Pos, Node::Varray, Node::Construct, etc.).

This distinction is a little inconvenient sometimes. For instance, to produce correct positions for record declarations, we need to know the position of the closing right brace. It would be wasteful to allocate a position for EVERY right brace, but with the way positions are currently tracked for tokens, we don't have much of a choice.

However since TokenKind is only one byte, and Node has an eight-byte data region, there are seven bytes available for us to store additional data without allocating. It's not enough space for a Pos, but it is enough space for a byte offset to the start of the token. For tokens with a fixed width (like right braces or keywords), the start offset is all we need to construct a Pos. 7 bytes is not quite as much as our normal storage for byte offsets (usize = 8 bytes), but it won't be a problem as long as no one attempts to parse a file larger than 67 million gigabytes, which doesn't seem like a scenario we need to concern ourselves with.

Reviewed By: shiqicao

Differential Revision: D24577864

fbshipit-source-id: cd3028717fb5f38324925e3ea07b8439506d3586