Add bytecodes for marking and tagging
Summary:
It's critical that marking and tagging happen in place whenever possible. If that wasn't true, we'd have to create (deep) copies of many core arrays.

However, an HHVM builtin that doesn't take its parameters as inout must never modify them. We rely on this assumption in refcount opts to remove refcounting around almost all builtin calls. This optimization breaks the marking and tagging builtins (as I can readily demonstrate on an example like the one in the new test). Using inout for these builtins is also not natural - among other things, we use them in constant locations like static class constants, which must get folded.

The simplest fix here is to replace these builtins with pseudo-functions that compile to bytecodes. The new bytecodes can easily operate on their inputs in place. Switching to bytecodes has a few other benefits, too. First, we were calling the builtins in a few other places in the parser - using bytecodes in these places is much more natural. Second, it's generally easier to optimize bytecodes than builtins. Here's a list of optimizations this diff already includes:

1. In HHBBC, if the input $value and $recursive are known, const-fold marking and unmarking.
1. In HHVM, skip tag_provenance_here in non-arrprov mode. (No use in HHBBC since we always build in arrprov mode).
1. In HHVM, the output of a marking or tagging call has the same type as the input.

The last of these optimizations is a new one. It'll also be easier to make further optimizations - for example, optimizing TagProvenanceHere in a non-SkipFrame function is now trivial. (Is that a case that matters?)

Reviewed By: ricklavoie

Differential Revision: D24584632

fbshipit-source-id: 8c99907a54b449e49824963bcb0321cb350cf6da