debug/dwarf: properly resolve line numbers for prologues
Summary:
When JitNoGdb is false, we try to record debug info for translations,
including line numbers. This logic was incorrect for prologues, as it tried to
obtain SrcKey::pc() and then resolving the offset of the bytecode instruction
to a line number. The problem is that prologues do not have any offsets.

Clean this up and use the SrcKey::lineNumber() API instead.

Reviewed By: oulgen

Differential Revision: D24602198

fbshipit-source-id: a5253b6eefb39d58f6a1c66ee2d29ab687854de4