Use two different implementations for FuncId based on LowPtr
Summary:
This diff adds the ability to back FuncId by address of LowPtr<const Func> when LowPtr is enabled. The addresses are equally unique as FuncId incremental numbers. This change enabled FuncId to Func* conversion to be instant, no lookups or anything.

With this change,
1) during live translations, SrcKey's funcid to func lookup is faster
2) s_funcVec is dead in LowPtr mode

This should give memory and perf wins.

On the next diff, I'll convert ActRec to always use FuncId instead of mix of FuncId and Lowptr<Func>.

Reviewed By: jano

Differential Revision: D24542896

fbshipit-source-id: 4c16ea2f93853c31adc7d49dd1bbcf800dd85c72